+++ app/change-distance/page.tsx
@@ -1,33 +1,20 @@
 'use client';
 
-import { useQuery } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
 import { useEffect, useState } from 'react';
 
 import { LoadingArea } from '@/components/atoms';
-import { BackButton, Dialog, HeaderBar } from '@/components/molecules';
-import SettingCalendar from '@/features/setting/components/organisms/setting-calendar';
-
-import { MemberProps } from '../api/member/route';
+import { BackButton, HeaderBar } from '@/components/molecules';
+import { useMemberData } from '@/features/setting/apis';
+import { SettingCalendar } from '@/features/setting/components';
+import { useSaveDialogHandler } from '@/features/setting/hooks';
 
 export default function Page() {
-  const fetchIdData = async (): Promise<MemberProps> => {
-    const response = await fetch(`/api/member`);
-    if (!response.ok) {
-      throw new Error('Failed to fetch profile data');
-    }
-    return (await response.json()) as MemberProps;
-  };
-
-  const { data, error } = useQuery<MemberProps>({
-    queryKey: ['data'],
-    queryFn: fetchIdData,
-  });
-
   const router = useRouter();
+  const { data, error } = useMemberData();
+  const { openSaveModal } = useSaveDialogHandler();
 
   const [selectedDistance, setSelectedDistance] = useState<number | null>(null);
-  const [isLogoutDialogOpen, setIsLogoutDialogOpen] = useState(false);
 
   useEffect(() => {
     if (data) {
@@ -36,7 +23,7 @@ export default function Page() {
   }, [data]);
 
   if (!data) return <LoadingArea />;
-  if (error) return <div>An error occurred: {error.message}</div>;
+  if (error) return console.log(error);
 
   const handleDistanceChange = (distance: number) => {
     setSelectedDistance(distance);
@@ -44,19 +31,10 @@ export default function Page() {
 
   const backButton = () => {
     if (data.data.goal !== selectedDistance) {
-      setIsLogoutDialogOpen(true);
+      openSaveModal();
     } else {
       router.push('/setting');
     }
-  };
-
-  const closeLogoutDialog = () => {
-    setIsLogoutDialogOpen(false);
-  };
-
-  const handleNavigateWithoutSaving = () => {
-    setIsLogoutDialogOpen(false);
-    router.push('/setting');
   };
 
   return (
@@ -72,24 +50,6 @@ export default function Page() {
         selectedDistance={selectedDistance}
         onDistanceChange={handleDistanceChange}
       />
-      {isLogoutDialogOpen && (
-        <Dialog
-          title="변경한 내용을 저장하지 않고 나갈까요?"
-          isOpen={isLogoutDialogOpen}
-          onClose={closeLogoutDialog}
-          buttons={{
-            confirm: {
-              text: '취소',
-              onClick: closeLogoutDialog,
-            },
-            cancel: {
-              text: '나가기',
-              onClick: handleNavigateWithoutSaving,
-            },
-          }}
-          isDim={true}
-        />
-      )}
     </div>
   );
 }
